Createt new Router and Endpoints
