+++ diabeticat-backend/DiabeticatAPI/API/Pet.py
@@ -1,0 +1,6 @@
+from fastapi import APIRouter
+
+bye_router = APIRouter()
+@bye_router.get('/bye')
+def bye():
+    return {"Schönen":"Tag Noch"}
