+++ diabeticat-backend/DiabeticatAPI/API/User.py
@@ -1,0 +1,6 @@
+from fastapi import APIRouter
+
+hello_router = APIRouter()
+@hello_router.get('/hello')
+def hello():
+    return {"Ping":"Pong"}
